+++ lib/schleuder/subscription.rb
@@ -21,8 +21,10 @@ module Schleuder
     end
 
     def fingerprint=(arg)
-      # Strip whitespace from incoming arg.
-      write_attribute(:fingerprint, arg.to_s.gsub(/\s*/, '').chomp)
+      # Allow input to contain whitespace and '0x'-prefix, but don't store it
+      # into the DB.
+      value = arg.to_s.gsub(/\s*/, '').gsub(/^0x/, '').chomp
+      write_attribute(:fingerprint, value)
     end
 
     def key
@@ -69,7 +71,7 @@ module Schleuder
     def ensure_headers(mail)
       mail.to = self.email
       mail.from = self.list.email
-      mail.return_path = self.list.bounce_address
+      mail.sender = self.list.bounce_address
       mail
     end