+++ CHANGELOG.md
@@ -3,6 +3,38 @@ Change Log
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.1.0] / 2017-05-21
+
+### Added
+
+* `X-GET-LOGFILE`. Sends you the logfile of the list.
+* `X-ATTACH-LISTKEY`. Attaches the list's key to a message. Useful in combination with `X-RESEND`.
+* `X-GET-VERSION`. Sends you the version of Schleuder that is running your list.
+* API-endpoint to trigger sending the list's key to all subscriptions.
+
+### Changed
+
+* Don't write errors of list-plugins into the list of pseudo-headers. List-plugins must handle errors on their own.
+* Allow request-plugins to return attachments.
+* Fix x-get-key for multiple keys per match, and attach the resulting keys.
+* Tolerate 0x-prefix on input for fingerprints of subscriptions.
+* Tolerate spaces on input for fingerprints in keywords.
+* `X-GET-KEY` returns keys as attachments now.
+* `X-SIGN-THIS` returns attachments now, too.
+* The texts that describe the forwarded automated messages now reflect that not all of those were bounces.
+* Use single SQL-query instead of five, in select-statement in postfix/schleuder_sqlite.cf.
+* Use sender() to specify the return-address, instead of setting a Return-Path.
+
+### Fixed
+
+* Make `public_footer` appear at the bottom of messages, not at the top.
+* Remove excessive empty lines in output of refresh-keys.
+* Amended list of dependencies in README.
+* Fix `X-GET-KEY` for multiple keys per match.
+* Also report if a key-import didn't change a present key.
+* Fix bounce-address in postfix/schleuder_sqlite.cf.
+
+
 ## [3.0.4] / 2017-04-15
 
 ### Changed